requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pandas>=2.0.0
openpyxl>=3.1.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.40.0
azure-identity>=1.15.0
azure-mgmt-network>=25.0.0
urllib3>=2.0.0
